pyproject.toml
[tool.poetry]
name = "redis-data-transfer"
version = "0.2.0"
description = "Transfer data between a redis instances or clusters"
authors = ["EDITED Devs <dev@edited.com>"]

[tool.poetry.dependencies]
python = "^3.7"
redis-py-cluster = "^2.0.0"
setproctitle = "^1.1.10"

[tool.poetry.dev-dependencies]
ipython = ""

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
redis-data-transfer = "redis_data_transfer:main"
